fixes in data type of Age field

+++ services/contact/internal/repository/storage/postgres/dao/contact.go
@@ -17,8 +17,8 @@ type Contact struct {
 	Surname    string `db:"surname"`
 	Patronymic string `db:"patronymic"`
 
-	Age    uint64 `db:"age"`
-	Gender uint8  `db:"gender"`
+	Age    uint8 `db:"age"`
+	Gender uint8 `db:"gender"`
 }
 
 var CreateColumnContact = []string{
